% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_div_factorizations.R
\name{add_factor}
\alias{add_factor}
\title{Helper function that adds one factor to the flash object.}
\usage{
add_factor(dat, loading, fl, prior, Fprior, allfixed = FALSE)
}
\arguments{
\item{dat}{the data matrix}

\item{loading}{binary initial loading to specify the sparsity pattern in the loading}

\item{fl}{flash object to add the factor to}

\item{prior}{prior for loadings}

\item{Fprior}{prior for the factors. Defaults to a normal prior.}

\item{allfixed}{Defaults to FALSE. If False, fixes only the loadings that are zero.
Otherwise, fixes the entire loading}
}
\value{
a new flash object with a new factor such that the associated loading matches the
sparsity pattern from loading
}
\description{
Helper function that adds one factor to the flash object.
}
