% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_div_factorizations.R
\name{div_cov_fit}
\alias{div_cov_fit}
\title{fits a divergence factorization to the covariance matrix
this function is less well tested}
\usage{
div_cov_fit(covmat, prior = ebnm::ebnm_point_laplace, Kmax = 1000)
}
\arguments{
\item{covmat}{the covariance matrix}

\item{prior}{prop for the loadings. Defaults to a point Laplace prior.}

\item{Kmax}{the maximum number of factors to add. Defaults to 1000}
}
\description{
fits a divergence factorization to the covariance matrix
this function is less well tested
}
