% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_div_factorizations.R
\name{drift_fit}
\alias{drift_fit}
\title{Fits a drift factorization to the data}
\usage{
drift_fit(
  dat,
  divprior = ebnm::ebnm_point_laplace,
  driftprior = ebnm::ebnm_point_exponential,
  Fprior = ebnm::ebnm_normal,
  Kmax = Inf,
  min_pve = 0,
  verbose.lvl = 0,
  eps = 0.01,
  labels = NULL
)
}
\arguments{
\item{dat}{data matrix. rows are samples, columns are features}

\item{divprior}{prior for intermediate divergence loadings. Defaults to a point Laplace prior.}

\item{driftprior}{prior for drift loadings. Defaults to a point exponential prior.}

\item{Fprior}{prior for the factors. Defaults to a normal prior.}

\item{Kmax}{maximum number of factors to add.}

\item{min_pve}{If the pve for a factor is less than this tolerance, the factor is rejected}

\item{verbose.lvl}{The level of verbosity of the function}

\item{eps}{Tolerance for nonzero values in the loadings.}

\item{labels}{Ground truth data labels for testing purposes}
}
\description{
Fits a drift factorization to the data
}
