% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileIO_plotting.R
\name{form_tree_from_file}
\alias{form_tree_from_file}
\title{Forms a tree object from a csv file}
\usage{
form_tree_from_file(filename)
}
\arguments{
\item{filename}{string containing the path to the csv file}
}
\value{
a tree vector with the following attributes.
csv: the data from the csv file
raw: selects columns Raw0:Raw499 in order to avoid the dimensionality-reduced data
matrix: the raw data cast to a matrix
dimred: the tsne dimensionality reducition from the csv file
counts: the "raw counts" matrix, which is computed as min(0,round(2**(tree$matrix)-1))
dataset: counts and matrix wraped for dynverse functions, with prior information that Row0 is the starting cell
trajectory: an empty vector for saving trajectory results
}
\description{
Forms a tree object from a csv file
}
