% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_div_factorizations.R
\name{get_divergence_factor}
\alias{get_divergence_factor}
\title{Helper function that returns the loading from one additional divergence factor}
\usage{
get_divergence_factor(dat, loading, fl, divprior, Fprior)
}
\arguments{
\item{dat}{the data matrix}

\item{loading}{binary initial loading to specify known sparsity pattern in the loading (e.g. parental sparsity)}

\item{fl}{flash object containing the current fit}

\item{divprior}{divergence prior for loadings}

\item{Fprior}{prior for the factors. Defaults to a normal prior.}
}
\value{
the new posterior loading for the additional divergence factor
}
\description{
Helper function that returns the loading from one additional divergence factor
}
