% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfsr_algorithm.R
\name{get_divergence_factor_lfsr}
\alias{get_divergence_factor_lfsr}
\title{Algorithm that uses the local false sign rate to construct a continuous form of the "drift"
factorization that is more robust to noise.
Helper function that returns the loading from one additional divergence factor.
Using the local false sign rate, it sets to zero those loadings for which we cannot be
confident of the sign.}
\usage{
get_divergence_factor_lfsr(dat, lfsr_tol, loading, fl, divprior, Fprior)
}
\arguments{
\item{dat}{the data matrix}

\item{lfsr_tol}{Tolerance for local false sign rate}

\item{loading}{binary initial loading to specify known sparsity pattern in the loading (e.g. parental sparsity)}

\item{fl}{flash object containing the current fit}

\item{divprior}{divergence prior for loadings}

\item{Fprior}{prior for the factors. Defaults to a normal prior.}
}
\value{
the new posterior loading for the additional divergence factor
}
\description{
Algorithm that uses the local false sign rate to construct a continuous form of the "drift"
factorization that is more robust to noise.
Helper function that returns the loading from one additional divergence factor.
Using the local false sign rate, it sets to zero those loadings for which we cannot be
confident of the sign.
}
