% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_div_factorizations.R
\name{get_sums_by_label}
\alias{get_sums_by_label}
\title{Functions to add drift and divergence factors
Helper function for debugging. Shows how many samples with each label type are activated on
a particular loading}
\usage{
get_sums_by_label(labels, loading)
}
\arguments{
\item{labels}{ground truth labels of data points}

\item{loading}{the loading in question}
}
\description{
Functions to add drift and divergence factors
Helper function for debugging. Shows how many samples with each label type are activated on
a particular loading
}
