% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfsr_algorithm.R
\name{lfsr_algorithm}
\alias{lfsr_algorithm}
\title{Fits a drift factorization to the data and returns the associated flashier object}
\usage{
lfsr_algorithm(
  dat,
  covar = FALSE,
  divprior = ebnm::ebnm_point_laplace,
  driftprior = ebnm::ebnm_point_exponential,
  Fprior = ebnm::ebnm_normal,
  Kmax = Inf,
  lfsr_tol = 0.001,
  min_pve = 0,
  verbose.lvl = 0,
  labels = NULL,
  allfixed = FALSE,
  S = 1e-12
)
}
\arguments{
\item{dat}{data matrix. rows are samples, columns are features}

\item{covar}{If true, fits to the covariance instead of the data matrix}

\item{divprior}{prior for intermediate divergence loadings. Defaults to a point Laplace prior.}

\item{driftprior}{prior for drift loadings. Defaults to a point exponential prior.}

\item{Fprior}{prior for the factors. Defaults to a normal prior.}

\item{Kmax}{maximum number of factors to add.}

\item{lfsr_tol}{Tolerance for local false sign rate}

\item{min_pve}{If the pve for a factor is less than this tolerance, the factor is rejected}

\item{verbose.lvl}{The level of verbosity of the function}

\item{labels}{Ground truth data labels for testing purposes}
}
\description{
Fits a drift factorization to the data and returns the associated flashier object
}
