% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileIO_plotting.R
\name{plot_nodetree_loadings}
\alias{plot_nodetree_loadings}
\title{For each loading, plots the dimensionality reduction of a nodetree
along with arrows along the branches, with data points colored according
to their value for the loading}
\usage{
plot_nodetree_loadings(
  tree,
  loadings,
  labels = tree$labels,
  leavesonly = FALSE
)
}
\arguments{
\item{tree}{tree object, like the output of form_tree_from_file}

\item{labels}{labels for the data used to color data in second plot}

\item{loading}{posterior mean of loadings, used to color the tree}
}
\description{
Also creates a second plot which shows the values of each loading,
with the data colored according to the grouping
}
