% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_div_factorizations.R
\name{run_methods}
\alias{run_methods}
\title{Run drift and divergence factorizations}
\usage{
run_methods(tree, Kmax, eps)
}
\arguments{
\item{tree}{a vector that is interpreted as a tree object; the output of form_tree_from_file from fileIO_plotting.R}

\item{Kmax}{parameter for the maximum number of factors to add}

\item{eps}{tolerance for when values are nonzero

This function does not return anything, but instead saves the results into tree$trajectory}
}
\description{
Run drift and divergence factorizations
}
